Accelerating and slowing the movement + scripts to convert tifs to png.
The function move_slow was added, and integrated into pattern2array.
It can be given margin in microns on which the piezo can slow down.
Also, the the array is cut to pieces by adding [-1,-1,-1] when the pattern is on edge
and is about to exceed the maxmal number of points for current array.

Two scripts are ment to be copied to /usr/loca/bin (on mac),
 and run from bash to convert to something easy to view.