Cleaned up the printing tools + inverted z axis.
The plan is to put actuall patterns in printing_patterns file, and helper
functions in prinitng_tools file. To create pattern run
printing_patterns.

Now the patterns will go form 100 to 0, which means they go from bottom
to top on the piezo station.

Fixed the problem with to fast movelent on even layers, still don't know
what was the cause of it.